backend/requirements.txt
praw==7.7.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
textblob==0.17.1
requests==2.31.0
schedule==1.2.0
